Documentation for BSE_SC.jl

Documentation for BSE_SC.jl

Index

Index

BSE_SC.OneToIndex_to_Freq

BSE_SC.aux_indices

BSE_SC.calc_χ₀

BSE_SC.improve_χ!

BSE_SC.read_gImp

BSE_SC.setup

BSE_SC.shell_indices

BSE_SC.ωn_to_ωi

BSE_SC.aux_indices — Method

aux_indices(ind_lst::Vector{CartesianIndex{2}}, ωi::Int, n_iω::Int, n_iν::Int, shift::Int)

Constructs two index lists, one for ν-ν' and one for ν+ν'+ω, from input index list ind_lst. These lists are used by improve_χ! interally.

BSE_SC.improve_χ! — Method

improve_χ!()

Improves asymptotics of χsp and χch. TODO: full documentation here.

BSE_SC.shell_indices — Method

shell_indices(Nν_full::Int, Nν_shell::Int)

Constructs 4 sets of indices for a square array of size Nν_full with a shell Nν_shell entries on each side in which asymptotic behaviour is assumed. Returns I_core, I_corner, I_top_bottom, I_left_right with indices of the corresponding regions.

BSE_SC.read_gImp — Method

read_gImp(input_file::String)

Read impurity Green's function from JLD2 file at path input_file.

BSE_SC.setup — Method

setup(input_file::String, N_shell::Int)

Setup calculation for data from file. This is used for testing from precomputed data. The file is expected to contain the following fields: β, U, gImp, grid_nBose, grid_nFermi, grid_shift, χDMFTsp, χDMFTch, χ_sp_asympt, χ_ch_asympt, χ_pp_asympt.

BSE_SC.OneToIndex_to_Freq — Method

OneToIndex_to_Freq(ωi::Int, νi::Int, n_iω::Int, n_iν::Int, shift::Int)

Translates Julia array indices for bosonic and fermionic axis (i.e. ωi ∈ [1,2n_iω+1], νi ∈ [1,2n_iν]) to Matsubara index (i.e. ωn ∈ [-n_iω,n_iω], νn ∈ [-n_iν -shiftωn/2, n_iν -shiftωn/2]). shift is used to center the non-asymptotic core of the vertex related functions in the center of the array.

BSE_SC.calc_χ₀ — Method

calc_χ₀(Gνω::OffsetArray, β::Float64, n_iω::Int, n_iν::Int, shift::Int)

Calculates the bubble term χ₀ from a given Green's function Gνω. This is mainly used for standalone testing and should generally be calculated independenty in the method using this package.

BSE_SC.ωn_to_ωi — Method

ωn_to_ωi(ωn::Int)

Translates Matsubara index (i.e. n ∈ [-N,N]) to Julia array index (i.e. i ∈ [1,2N+1]), assuming symmetry around 0!