Documentation for BSE_SC.jl

Documentation for BSE_SC.jl

Index

Index

BSE_SC.BSE_Asym_Helper

BSE_SC.BSE_SC_Helper

BSE_SC.OneToIndex_to_Freq

BSE_SC.aux_indices

BSE_SC.calc_λ0_impr

BSE_SC.calc_χλ_impr

BSE_SC.calc_χ₀

BSE_SC.improve_χ!

BSE_SC.read_gImp

BSE_SC.setup

BSE_SC.shell_indices

BSE_SC.χ₀_shell_sum

BSE_SC.χ₀_shell_sum_core

BSE_SC.ωn_to_ωi

BSE_SC.calc_λ0_impr — Method

calc_λ0_impr(type::Symbol, ωgrid::AbstractVector{Int},
             F::AbstractArray{ComplexF64,3}, χ₀::AbstractArray{ComplexF64,3}, 
             χ₀_asym::Array{ComplexF64,2}, γ::AbstractArray{ComplexF64,2}, 
             χ::AbstractArray{ComplexF64,1},
             U::Float64, β::Float64, h::BSE_Asym_Helper)

Calculates improved version of λ₀ = χ₀ ⋆ F. TODO: finish documenation and tests.

BSE_SC.calc_χλ_impr — Method

calc_χλ(type::Symbol, ωn::Int, χ::AbstractArray{ComplexF64,2}, χ₀::AbstractArray{ComplexF64,1}, U::Float64, β::Float64, χ₀_asym::Float64, h::BSE_Asym_Helper)

Calculates the physical susceptibility χ and triangular vertex λ in a given channel type=:sp or type=:ch using knowledge about the asymptotics of the full vertex and tails of the Green's function. χ₀_asym is the ω dependent asymptotic tail of χ₀ and can be calculated with χ₀_shell_sum.

TODO: optimize - test for useless allocations - define subroutine version, for map over ωn - fft for F_diag?

BSE_SC.improve_χ! — Method

improve_χ!(type::Symbol, ωi::Int, χr, χ₀, U, β, h::BSE_SC_Helper; 
           Nit=200, atol=1e-9)

Improves asymptotics of χr, given a channel type = :sp or :ch, using Eq. 12a 12b from DOI: 10.1103/PhysRevB.97.235.140 ωn specifies the index of the Matsubara frequency to use. χ₀ is the bubble term, U and β Hubbard-U and temperature. h is a helper struct, see BSE_SC_Helper. Additionally one can specify convergence parameters: Nit: maximum number of iterations atol: minimum change between total sum over χ between iterations.

BSE_SC.BSE_Asym_Helper — Type

BSE_Asym_Helper

Helper struct for efficient solution of the self consistency. For an example see also setup.

Fields

χsp_asympt : ω-asymptotic for the susceptibility in the spin-ph channel

χch_asympt : ω-asymptotic for the susceptibility in the charge-ph channel

χpp_asympt : ω-asymptotic for the susceptibility in the pp channel

Nν_shell : Number of Fermionic frequencies used for asymptotic extension.

I_core : Indices for the core (non-asymptotic) region.

I_asympt : Indices for the asymptotic region (union of Icorner, It, I_r).

ind1_list : ν-ν' indices in χsp_asympt and χch_asympt for all I_asympt

ind2_list : ν+ν'+ω indices in χpp_asympt for all I_asympt

shift : 1 or 0 depending on wheter the ν-frequencies are shfited by -ω/2

BSE_SC.BSE_SC_Helper — Type

BSE_SC_Helper

Helper struct for efficient solution of the self consistency. For an example see also setup.

Fields

χsp_asympt : ω-asymptotic for the susceptibility in the spin-ph channel

χch_asympt : ω-asymptotic for the susceptibility in the charge-ph channel

χpp_asympt : ω-asymptotic for the susceptibility in the pp channel

Fr : Temporary storage for the full vertex. Used internally to avoid repeated mallocs.

λr : Temporary storage for the fermion-boson-vertex. Used internally to avoid repeated mallocs.

Nν_shell : Number of Fermionic frequencies used for asymptotic extension.

I_core : Indices for the core (non-asymptotic) region.

I_corner : Indices for the corner (ν AND ν' outside core) region.

I_t : Indices for the core (ν outside core) region.

I_r : Indices for the core (ν' outside core) region.

I_asympt : Indices for the asymptotic region (union of Icorner, It, I_r).

ind1_list : ν-ν' indices in χsp_asympt and χch_asympt for all I_asympt

ind2_list : ν+ν'+ω indices in χpp_asympt for all I_asympt

shift : 1 or 0 depending on wheter the ν-frequencies are shfited by -ω/2

BSE_SC.aux_indices — Method

aux_indices(ind_lst::Vector{CartesianIndex{2}}, ωi::Int, n_iω::Int, n_iν::Int, shift::Int)

Constructs two index lists, one for ν-ν' and one for ν+ν'+ω, from input index list ind_lst. These lists are used by improve_χ! interally.

BSE_SC.shell_indices — Method

shell_indices(Nν_full::Int, Nν_shell::Int)

Constructs 4 sets of indices for a square array of size Nν_full with a shell Nν_shell entries on each side in which asymptotic behaviour is assumed. Returns I_core, I_corner, I_top_bottom, I_left_right with indices of the corresponding regions.

BSE_SC.χ₀_shell_sum — Method

χ₀_shell_sum(core::OffsetMatrix{ComplexF64}, ωn::Union{Int,AbstractVector{Int}}, c2::Float64, c3::Float64)

Calculates the asymptotic sum of ∑ₙ,ₗ χ₀(iωₘ,iνₙ,iνₗ') with n,l from n_iν to ∞. This is done by using the known first three tail coefficients of the Green's function in iνₙ and expanding around n → ∞. ωn can either be a single Matsubara index or an array of indices. The core region is precalculated using χ₀_shell_sum_core.

BSE_SC.χ₀_shell_sum_core — Method

χ₀_shell_sum_core(β::Float64, ω_ind_grid::AbstractVector{Int}, n_iν::Int, shift::Int)

Calculates the core region for use in χ₀_shell_sum.

BSE_SC.read_gImp — Method

read_gImp(input_file::String)

Read impurity Green's function from JLD2 file at path input_file.

BSE_SC.setup — Method

setup(input_file::String, N_shell::Int)

Setup calculation for data from file. This is used for testing from precomputed data. The file is expected to contain the following fields: β, U, gImp, grid_nBose, grid_nFermi, grid_shift, χDMFTsp, χDMFTch, χ_sp_asympt, χ_ch_asympt, χ_pp_asympt.

BSE_SC.OneToIndex_to_Freq — Method

OneToIndex_to_Freq(ωi::Int, νi::Int, n_iω::Int, n_iν::Int, shift::Int)

Translates Julia array indices for bosonic and fermionic axis (i.e. ωi ∈ [1,2n_iω+1], νi ∈ [1,2n_iν]) to Matsubara index (i.e. ωn ∈ [-n_iω,n_iω], νn ∈ [-n_iν -shiftωn/2, n_iν -shiftωn/2]). shift is used to center the non-asymptotic core of the vertex related functions in the center of the array.

BSE_SC.calc_χ₀ — Method

calc_χ₀(Gνω::OffsetArray, β::Float64, n_iω::Int, n_iν::Int, shift::Int)

Calculates the bubble term χ₀ from a given Green's function Gνω. This is mainly used for standalone testing and should generally be calculated independenty in the method using this package.

BSE_SC.ωn_to_ωi — Method

ωn_to_ωi(ωn::Int)

Translates Matsubara index (i.e. n ∈ [-N,N]) to Julia array index (i.e. i ∈ [1,2N+1]), assuming symmetry around 0!